+++ R/stylo.R
@@ -15,6 +15,9 @@ stylo = function(gui = TRUE,
              parsed.corpus = NULL,
              features = NULL,
              path = NULL,
+             metadata = NULL,
+             filename.column = "filename",
+             grouping.column = "author",
              corpus.dir = "corpus", ...) {
 
 
@@ -48,17 +51,6 @@ if(is.character(path) == TRUE & length(path) > 0) {
 
 if(is.character(corpus.dir) == FALSE | nchar(corpus.dir) == 0) {
   corpus.dir = "corpus"
-}
-
-
-# Choose directory:
-#
-# Just a few lines that allow users to choose the working directory if working
-# with the GUI.
-
-if(gui == TRUE & is.null(path)){
-  selected.path = tk_choose.dir(caption = "Select your working directory. It should a subdirectory called *corpus* ")
-  setwd(selected.path)
 }
 
 
@@ -534,6 +526,22 @@ if(corpus.exists == TRUE) {
 
 if(corpus.exists == FALSE) {
 
+  # Checking whether the required subdirectory exists, calling the choose directory dialogue if not.
+  if(file.exists(corpus.dir) == FALSE) {
+    selected.path = tk_choose.dir(caption = "Select your working directory. It should have a subdirectory called *corpus* ")
+    setwd(selected.path)
+  }
+  if(file.exists(corpus.dir) == FALSE) {
+          message("\n\n", "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n",
+              "Hey! The working directory should contain the subdirectory \"",
+              corpus.dir,"\"\n",
+              "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n\n")
+          # back to the original working directory
+          setwd(original.path)
+          # error message
+          stop("Corpus prepared incorrectly")
+  }
+  
   # Retrieving the names of texts.
   # It's possible to choose the files manually (choose an appropriate option!)
   if (interactive.files == TRUE) {
@@ -574,18 +582,7 @@ if(corpus.exists == FALSE) {
       }
   }
 
-
-  # Checking whether the required files and subdirectory exist
-    if(file.exists(corpus.dir) == FALSE) {
-      message("\n\n", "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n",
-          "Hey! The working directory should contain the subdirectory \"",
-          corpus.dir,"\"\n",
-          "!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n\n")
-      # back to the original working directory
-      setwd(original.path)
-      # error message
-      stop("Corpus prepared incorrectly")
-    }
+  # Checking whether the required files exist
     if(length(corpus.filenames) <2 & sampling != "normal.sampling")  {
       message("\n\n !!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!!\n",
           "Ho! The subdirectory \"",corpus.dir,"\" should contain at least
@@ -1124,13 +1121,17 @@ distance.table = as.matrix(distance.table)
 
 # #################################################
 # a tiny module for graph auto-coloring:
-# uses the function "assign.plot.colors()"
+# uses the functions "metadata.processing()"
+# and assign.plot.colors()"
 # #################################################
 
-names.of.texts = gsub("(\\.txt)||(\\.xml)||(\\.html)||(\\.htm)","",rownames(table.with.all.freqs))
+groups = process.metadata(metadata = metadata, 
+                          filenames = rownames(table.with.all.freqs),
+                          filename.column = filename.column,
+                          grouping.column = grouping.column)
 
 # using an appropriate function to assing colors to subsequent samples
-colors.of.pca.graph = assign.plot.colors(labels = names.of.texts,
+colors.of.pca.graph = assign.plot.colors(labels = groups,
                             col = colors.on.graphs, opacity = 1)