+++ NEWS.md
@@ -1,4 +1,15 @@
-## stylo news ##
+## 'stylo' news ##
+
+
+### version 0.7.0, 2019/01/22
+  * Unicode (UTF-8) made the default encoding, also for Windows
+
+
+### version 0.6.9, 2019/01/20
+  * check.encoding() and change.encoding() introduced
+  * GUI allows for changing the working directory with one click
+  * metadata handling through a dedicated variable
+  * {Steffen Pielström joins!}
 
 
 ### version 0.6.8, 2018/06/14